requirements.txt
Flask==2.3.2
pyjwt==2.6.0
requests==2.31.0
